% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stats.R
\name{basic_stats}
\alias{basic_stats}
\title{Basic summary stats by group}
\usage{
basic_stats(data, x, by, na.rm = TRUE, ...)
}
\arguments{
\item{data}{a data frame or tibble containing \code{x} and \code{by}.}

\item{x}{a numeric variable in \code{data} to compute summary statistics for.}

\item{by}{a factor or character variable in \code{data} by which to compute summary statistics.}

\item{na.rm}{a logical evaluation to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped.}

\item{...}{for further parameters to modify the calculation and formatting of the summary statistics.}
}
\value{
a tibble with mean, median, range, and sample size calculations by levels of your chosen grouping variable.
}
\description{
Presents mean, median, range, and sample size by group in an easy-to-read tibble.
}
\details{
basic_stats converts data frames to tibbles. It also converts any character variable with fewer than 20 unique values to a factor variable. If the character variable contains more than 20 unique values, an error will be issued. This constraint can be overridden by converting your grouping variable to a factor before passing it into the function.
}
\examples{
# Basic usage:
basic_stats(iris, Sepal.Width, Species)
# Produces summary stats for Sepal.Width by levels
# of Species from the iris dataset.

# Advanced usage:
basic_stats(iris, Sepal.Width, Species, na.rm = FALSE, trim = 0.1)
# Modifies the previous example by retaining
# null values and trimming the 10\% most extreme values.
}
