% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions_dataframe.R
\name{dimensions}
\alias{dimensions}
\title{Dimensions in a dataframe}
\usage{
dimensions(x)
}
\arguments{
\item{A}{dataframe}
}
\value{
The rows and columns of a dataframe
}
\description{
Takes in the dataframe and prints the number or rows and columns
}
