% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAE.R
\name{MAE}
\alias{MAE}
\title{Calculate mean absolute error}
\usage{
MAE(obs, pred)
}
\arguments{
\item{obs}{Observed values.}

\item{pred}{Predicted values.}
}
\value{
The mean absolute error.
}
\description{
Calculate mean absolute error
}
