% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean_df}
\alias{mean_df}
\title{Mean of every attribute in dataframe}
\usage{
mean_df(x)
}
\arguments{
\item{A}{dataframe}
}
\value{
The column name and its mean
}
\description{
Takes in the dataframe as input and print out the column name and its mean value
If the column is not a numeric attribute then it throws NA value for  mean of that column
}
