% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediandf.R
\name{median_df}
\alias{median_df}
\title{Median of every attribute in a dataframe}
\usage{
median_df(x)
}
\arguments{
\item{A}{dataframe}
}
\value{
The column name and its median
}
\description{
Takes in the dataframe as input and print out the column name and its median value
If the column is not a numeric attribute then it throws NA value for median of that column
}
