% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max.R
\name{min_max_df}
\alias{min_max_df}
\title{Minimum and Maximum of every attribute in a dataframe}
\usage{
min_max_df(x)
}
\arguments{
\item{A}{dataframe}
}
\value{
The column name and its Minimum, maximum value
}
\description{
Takes in the dataframe as input and print out the column name and its Minimum, maximum value
If the column is not a numeric attribute then it throws NA value for value of that column
}
