% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode_df}
\alias{mode_df}
\title{Mode of every attribute in dataframe}
\usage{
mode_df(x)
}
\arguments{
\item{A}{dataframe}
}
\value{
The column name and its mode
}
\description{
Takes in the dataframe as input and print out the column name and its mode value
If the column is not a numeric attribute then it throws NA value for  mode of that column
}
