% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE.R
\name{RMSE}
\alias{RMSE}
\title{Calculate root mean squared error}
\usage{
RMSE(obs, pred)
}
\arguments{
\item{obs}{Observed values.}

\item{pred}{Predicted values.}
}
\value{
The root mean squared error.
}
\description{
Calculate root mean squared error
}
