% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_deviation.R
\name{std_dev_df}
\alias{std_dev_df}
\title{Standard Deviation of every attribute in a dataframe}
\usage{
std_dev_df(x)
}
\arguments{
\item{A}{dataframe}
}
\value{
The column name and its Standard Deviation
}
\description{
Takes in the dataframe as input and print out the column name and its Standard Deviation value
If the column is not a numeric attribute then it throws NA value for Standard Deviation of that column
}
