% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variance.R
\name{var_df}
\alias{var_df}
\title{Variance of every attribute in a dataframe}
\usage{
var_df(x)
}
\arguments{
\item{A}{dataframe}
}
\value{
The column name and its Variance
}
\description{
Takes in the dataframe as input and print out the column name and its Variance value
If the column is not a numeric attribute then it throws NA value for Variance of that column
}
