% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stats.R
\name{basic_stats}
\alias{basic_stats}
\title{Basic summary stats by group}
\usage{
basic_stats(data, x, by, na.rm = TRUE, ...)
}
\arguments{
\item{data}{a data frame or tibble containing \code{x} and \code{by}.}

\item{x}{a numeric variable in the dataset to compute summary statistics for.}

\item{by}{a categorical variable in the dataset by which to compute summary statistics.}

\item{na.rm}{a logical evaluation to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped.}

\item{...}{for further arguments to modify the calculation and formatting of the summary statistics.}
}
\value{
a tibble with mean, median, range, and sample size calculations by levels of your chosen grouping variable.
}
\description{
Presents mean, median, range, and sample size by group in an easy-to-read tibble.
}
\examples{
# Basic usage:
basic_stats(iris, Sepal.Width, Species)
# Advanced usage:
basic_stats(iris, Sepal.Width, Species, na.rm = FALSE, trim = 0.1)
}
